Move nav bar icon button to a new component
